0000 Fee Evaluation Exploration  * Added a simple expression evaluator     * This evaluator is simple arithmetic (add, subtract, multiply, divide, and parenthetic expressions) and long values only.  * Added parameter expansion in the FeeCalculatorImpl class, it's not fast, but it's simple and correct.  * Added a test for the expression evaluator  * Cleaned up a few items the IDE complained about

+++ fee-calculation/src/main/java/com/hedera/fee/calculation/FeeCalculator.java
@@ -3,7 +3,6 @@ package com.hedera.fee.calculation;
 import com.hedera.fee.calculation.impl.FeeCalculatorImpl;
 import com.hedera.fee.calculation.model.OperationMetadata;
 import com.hedera.fee.calculation.model.ServiceMetadata;
-import java.math.BigDecimal;
 import java.util.Map;
 import java.util.Set;
 
@@ -25,5 +24,5 @@ public sealed interface FeeCalculator permits FeeCalculatorImpl {
    * @param input the input values
    * @return the result of the calculation
    */
-  BigDecimal calculate(OperationMetadata operation, final Map<String, BigDecimal> input);
+  long calculate(OperationMetadata operation, final Map<String, Long> input);
 }
